+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [1.0.0] - 2020-06-23
+## [1.0.2] - 2020-07-16
 ### Added
-- Added bundles loaders
+- In edit mode in Unity Editor sprites and textures that have bundle loaders with specified sprite asset from a bundle will set their respective renderers with these sprites to enhance visualization while working with prefabs
+
 ## [1.0.1] - 2020-07-08
 ### Added
 - Added bundle loaders for SpriteRenderer component
 ### Refactored
 - moved bundle loading logic to base classes 
+
+## [1.0.0] - 2020-06-23
+### Added
+- Added bundles loaders